requirements.txt
file: requirements.txtPython dependencies with pinned versions for reproducibilityPython 3.10+ requiredCore data processingpandas==2.0.3numpy==1.24.3Web scraping (optional, templates only)requests==2.31.0beautifulsoup4==4.12.2scrapy==2.11.0lxml==4.9.3NLP and MLspacy==3.6.1en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whlsentence-transformers==2.2.2transformers==4.33.2torch==2.0.1Machine learning utilitiesscikit-learn==1.3.0joblib==1.3.2Network analysisnetworkx==3.1Visualizationmatplotlib==3.7.2seaborn==0.12.2Time series analysisstatsmodels==0.14.0prophet==1.1.4Association rule miningmlxtend==0.22.0ConfigurationPyYAML==6.0.1Testingpytest==7.4.2pytest-cov==4.1.0Utilitiestqdm==4.66.1python-dateutil==2.8.2pytz==2023.3Optional document processingpython-docx==0.8.11openpyxl==3.1.2Code quality (optional)black==23.9.1flake8==6.1.0mypy==1.5.1